スタート時間を初期化

+++ philo/check.c
@@ -6,7 +6,7 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/26 11:37:04 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2025/11/26 11:39:02 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2025/11/27 10:43:52 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ t_bool	is_died(t_atomic_bool *is_died)
 	return (res);
 }
 
-t_bool	is_timeout_died(const long last_eat_time, const int time_to_died)
+t_bool	is_timeout_died(const  long long last_eat_time, const int time_to_died)
 {
 	return (get_current_mstime() - last_eat_time > time_to_died);
 }
