+++ philo/general.c
@@ -6,7 +6,7 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 10:40:36 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2025/11/22 14:04:40 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2025/11/27 10:26:23 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ t_bool ft_is_digit(int ch)
 	return (ch >= '0' && ch <= '9');
 }
 
-long	ft_atol(const char *nptr)
+long long	ft_atol(const char *nptr)
 {
 	int			sign;
 	long long	res;
