+++ philo/get.c
@@ -6,23 +6,23 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/23 14:30:17 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2025/11/26 12:05:18 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2025/11/27 10:28:48 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long	get_current_mstime(void)
+long long	get_current_mstime(void)
 {
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return ((long)(tv.tv_sec * 1000) + (long)(tv.tv_usec / 1000));
+	return ((long long)(tv.tv_sec * 1000) + (long long)(tv.tv_usec / 1000));
 }
 
-long	get_last_eat_time(t_atomic_long *last_eat_time)
+long long	get_last_eat_time(t_atomic_long *last_eat_time)
 {
-	long	res;
+	long long	res;
 
 	pthread_mutex_lock(&last_eat_time->lock);
 	res = last_eat_time->value;
@@ -30,7 +30,7 @@ long	get_last_eat_time(t_atomic_long *last_eat_time)
 	return (res);
 }
 
-long	get_time_stamp(const long start_time)
+long long	get_time_stamp(const long long start_time)
 {
 	return (get_current_mstime() - start_time);
 }
