+++ philo/parse.c
@@ -6,28 +6,28 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 13:09:52 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2025/11/24 11:19:36 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2025/11/27 10:24:52 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-static t_bool	is_zero(const long n)
+static t_bool	is_zero(const long long n)
 {
 	return (n == 0);
 }
 
-static t_bool	is_overflow(const long n)
+static t_bool	is_overflow(const long long n)
 {
 	return (n > INT_MAX || n < INT_MIN);
 }
 
-static t_bool	is_negative(const long n)
+static t_bool	is_negative(const long long n)
 {
 	return (n < 0);
 }
 
-static t_bool	parse_check(const long *arr, size_t size)
+static t_bool	parse_check(const long long *arr, size_t size)
 {
 	size_t	i;
 
@@ -48,7 +48,7 @@ static t_bool	parse_check(const long *arr, size_t size)
 t_bool	parse(char **argv, t_philo_config *p_conf)
 {
 	size_t i;
-	long int arr[CONF_SIZE];
+	long long int arr[CONF_SIZE];
 	int *p;
 
 	i = 0;
